move indent.ts over to paragraph and add a #indent method
